IMPROVE speed, delay timers, drawing function
"Upcycling" the interpreter by doing more in a frame, just a simple [for loop] in a main loop.
Seems like 30 or 60Hz isn't enough for some games/responsiveness. TODO: make it configurable.
Delay timers got an additional limiter % to compensate the main clock changes .

A bit of refactoring, plus one module added for the specific app config and utils.